the_pass/requirements.txt
apt-xapian-index==0.49
asgiref==3.7.2
blinker==1.4
chardet==4.0.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.1
Django==5.0
django-filter==23.5
djangorestframework==3.14.0
drf-writable-nested==0.7.0
fuse-python==1.0.2
gpg==1.16.0
httplib2==0.20.2
importlib-metadata==4.6.4
jeepney==0.7.1
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
Markdown==3.5.1
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.0
olefile==0.46
pexpect==4.8.0
Pillow==10.1.0
psycopg==3.1.16
ptyprocess==0.7.0
pycairo==1.20.1
pycups==2.0.1
PyGObject==3.42.1
PyJWT==2.3.0
pylibacl==0.6.0
pyparsing==2.4.7
PyQt5==5.15.6
PyQt5-sip==12.9.1
python-apt==2.4.0+ubuntu2
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.0
pytz==2023.3.post1
pyxattr==0.7.2
PyYAML==5.4.1
reportlab==3.6.8
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
six==1.16.0
sqlparse==0.4.4
systemd-python==234
tornado==6.1
typing_extensions==4.9.0
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
usb-creator==0.3.7
vboxapi==1.0
wadllib==1.3.6
xkit==0.0.0
zipp==1.0.0
